Moved accepting invitation code into new invitations module
- Following the logic that controller should only do validation/permission checks/response formatting moving out the rest of code to be reused in new v2 controller
- Feels like the whole acceptInvitation method belongs to invitations
controller rather then 'authentication' but not messing with that at
this moment
