+++ core/server/services/invitations/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    get accept() {
+        return require('./accept');
+    }
+};
